Rename eiquadprog.hpp to eiquadprog.h. This brings the filename in line with the google-styleguide.
