+++ drake/systems/plants/eiquadprog.h
